requirements.txt
flask==3.0.0
python-dotenv>=1.0.0
gunicorn==21.2.0
python-chess
numpy>=1.24
keras>=3.0.0
torch>=2.0.0
mistralai>=1.0.0